Rename adj_pred_probs to adjust_pred_probs. Move get_confident_thresholds() to count.py module.
