+++ cleanlab/utils/label_quality_utils.py
@@ -17,38 +17,10 @@
 """Helper functions for computing label quality scores"""
 
 import numpy as np
+from cleanlab.count import get_confident_thresholds
 
 
-def get_confident_thresholds(labels: np.array, pred_probs: np.array) -> np.array:
-    """Returns expected (average) "self-confidence" for each class.
-
-    The confident class threshold for a class j is the expected (average) "self-confidence" for class j.
-
-    Parameters
-    ----------
-    labels : np.array
-      A discrete vector of noisy labels, i.e. some labels may be erroneous.
-      *Format requirements*: for dataset with K classes, labels must be in {0,1,...,K-1}.
-
-    pred_probs : np.array (shape (N, K))
-      P(label=k|x) is a matrix with K model-predicted probabilities.
-      Each row of this matrix corresponds to an example x and contains the model-predicted
-      probabilities that x belongs to each possible class.
-      The columns must be ordered such that these probabilities correspond to class 0,1,2,...
-      `pred_probs` should have been computed using 3 (or higher) fold cross-validation.
-
-    Returns
-    -------
-    confident_thresholds : np.array (shape (K,))
-
-    """
-    confident_thresholds = np.array(
-        [np.mean(pred_probs[:, k][labels == k]) for k in range(pred_probs.shape[1])]
-    )
-    return confident_thresholds
-
-
-def subtract_confident_thresholds(labels: np.array, pred_probs: np.array) -> np.array:
+def _subtract_confident_thresholds(labels: np.array, pred_probs: np.array) -> np.array:
     """Returns adjusted predicted probabilities by subtracting the class confident thresholds and renormalizing.
 
     The confident class threshold for a class j is the expected (average) "self-confidence" for class j.
@@ -58,10 +30,10 @@ def subtract_confident_thresholds(labels: np.array, pred_probs: np.array) -> np.
     Parameters
     ----------
     labels : np.array
-      Labels in the same format expected by the `get_confident_thresholds()` method.
+      Labels in the same format expected by the `cleanlab.count.get_confident_thresholds()` method.
 
     pred_probs : np.array (shape (N, K))
-      Predicted-probabilities in the same format expected by the `get_confident_thresholds()` method.
+      Predicted-probabilities in the same format expected by the `cleanlab.count.get_confident_thresholds()` method.
 
     Returns
     -------
